+++ src/Lights/Group.cpp
@@ -86,7 +86,7 @@ void Group::setIntensity(unsigned char targetI,  float fadeTime, int elementInde
     }
 }
 
-void Group::makeElementsFollow(int Amin, int Amax, signalState signal, int freq, int phase, float randomComponent, int elementIndex)
+void Group::makeElementsFollow(int Amin, int Amax, signalState signal, float freq, int phase, float randomComponent, int elementIndex)
 {
     if (elementIndex == -1){    // Everyone
         for(int i = 0; i < elements.size(); i++)
