+++ src/Lights/Group.h
@@ -45,7 +45,7 @@ public:
     void setColor(ofColor color, float fadeTime, int elementIndex);
     void setIntensity(unsigned char targetI, float fadeTime, int elementIndex);
     
-    void makeElementsFollow(int Amin, int Amax, signalState signal, int freq, int phase, float randomComponent, int elementIndex);
+    void makeElementsFollow(int Amin, int Amax, signalState signal, float freq, int phase, float randomComponent, int elementIndex);
 
     
 };
